+++ lib/components/fo_dropdown_list/fo_dropdown_list_component.dart
@@ -41,6 +41,9 @@ class FoDropdownListComponent<T> implements AfterChanges, OnDestroy {
   bool constrainToViewPort = true;
 
   @Input()
+  bool showCategoryLabels = true;
+
+  @Input()
   bool materialIcons = true;
 
   @Input()